windows fix for http://bugzilla.zeroc.com/bugzilla/show_bug.cgi?id=1274

+++ cpp/src/IceGrid/Client.cpp
@@ -282,7 +282,7 @@ Client::run(int argc, char* argv[])
 		    
 		session = IceGrid::AdminSessionPrx::uncheckedCast(router->createSession(id, password));
 	    }
-	    timeout = router->getSessionTimeout();
+	    timeout = static_cast<int>(router->getSessionTimeout());
 	}
 	else
 	{
